Refactor game object arrays to ArrayLists and add platform-fall feature
In this commit, the static GameObject arrays are replaced with ArrayLists for flexibility. This allows gameObjects to be dynamically added, offering enhanced game dynamics. Additionally, we introduced a "fallable" attribute to GameObject. Objects with this attribute will leave the field when player steps on them. Implemented the deletion of fallen platforms from the gameObjects list, which optimizes active gameObjects handling by reducing computations for obsolete platforms. The GameObject's shrinking feature is also introduced to visibly represent the platforms dropping out of gameplay when stepped on by the player.