Added platform classes and refactored player and world
In this commit, new classes for handling different types of platforms have been added. The platforms have specific behaviors, such as Ground, Ramp, Brick , Trap and Moving Belt. In addition to this, the Player and World class have been adjusted to incorporate these new changes. This has been done to make the game base more modular, scalable and prepare it for new platform types in the future.