Add collectibles feature to game
A new feature to collect and score objects has been introduced. The feature includes the addition of a Collectible class to define the objects that can be collected in a game. Within the Player class key modifications are done to count score, keys and to interact with the collectibles. Modifications are also done in the World class to add and draw collectibles. Finally, in the DrawingThread class, checkCollectibles function is called to check and gather collectibles.

In nutshell, these changes will enhance the game play by introducing the concepts of scoring and collecting items.