Add game state logic, save and load score to file
Added game state check in both `Player` and `World` class. In `Player` class we now check if lives are depleted and in `World` class we check if the player has hit the finish line, thereby ending the game. A new file, `score.txt` in `data` directory has been introduced to store highest score played and the score is retrieved everytime the game starts. The `World` class is also responsible for saving the score if a higher score is achieved during current play of game.