"Added 'die' sound effect and improved player score calculation
Introduced a new 'die' sound effect in Constants.java and Music.java, which is played in Player.java when the player dies either from damage, lack of air, or falling. The score calculation following a death event in Player.java has also been adjusted to be multiplied by the number of keys. These changes enhance player immersion and add a more rewarding dynamic to the score system."